Rename 'glossary' to 'acronyms'
They are not the same thing, and the current heading calls it "Acronyms" already.
Also the "Empfehlungen zu Abschlussarbeiten" mentions an "Abkürzungsverzeichnis".
It also mentions a "Stichwortverzeichnis" (which might be the same as a 'glossary')
but says this should be rarely needed.

If necessary can consider adding a separate glossary or similar to the template in the future.